pyproject.toml
[project]
name = "askcli"
version = "0.1.0"
description = "A CLI tool for intelligent Q&A with web search capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "trafilatura>=1.6.0",
    "rich>=13.0.0",
    "groq>=0.4.0",
    "google-generativeai>=0.3.0",
    "toml>=0.10.0",
    "ddgs>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ask = "askcli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
